Improved packager even further
Fixed bug in filesystem where bad directory entries were leaking
through, causing random data loss. In the same vein, added the ability
for files to be passed directly to the overlay filesystem as readonly,
which allows _only_ the files named in a buildrule and the outputs of
dependencies to be loaded into the fake root. Also fixed a re-build bug
with > being swapped with <.